Unbreak OSX builds
Something appears to have changed with Travis, which is causing our OSX
builds to fail now. The issue seems to be related to RVM to some extent,
but figuring out exactly what it is is probably not worth the effort.
This patch works around by using RVM's `rvmsudo` when it detects RVM.