Bundle OpenSSH
We'd like to start using UNIX port-forwards on the server side for DB
tunnels, but that feature shipped only two years ago (...), so users are
likely to have SSH versions that don't support it (e.g. Trusty and
Precise).

This solves that by including OpenSSH in our binary, and putting it on
the `PATH` when we run the Aptible CLI (I've also included a tweak I had
laying around to unset `APPBUNDLER_ALLOW_RVM` when running the Aptible
CLI).