Git fetch on build
Handle another edge case with git sources, where simply cleaning isn't
enough, because Omnibus's Software fetcher is perfectly happy with using
the git SHA that corresponds to the version we are trying to build (it
uses `git ls-remote` for that) without actually fetching that tag,
whereas Omnibus's build version engine actually wants the tag locally to
run `git describe`.

As a result, merely running fetcher#clean won't download the tag, and
the resulting package version won't be the right tag (it'll be the last
tag we pulled plus a number of commit, i.e. a pre-release).

So, we now also fetch the git source..!