Windows: use a .exe entrypoint rather than .bat
We currently use a aptible.bat script as a shim on Windows in order to
set the environment variables we need, which is similar to what we do on
OSX / Linux where we use a /bin/sh script.

Unfortunately, while this works fine on OSX / Linux (where we have
access to `exec`), it's not so great on Windows:

- Calling a batch file from another batch file is a bad idea.
- It's unclear if you can even call a batch file from another program.
- I'm not sure what we're doing with the exit code, if anything.

And most importantly, whenever you CTRL+C out of the program, it
triggers an ugly and OCD-inducing "Terminate batch job (Y/N)" prompt,
and god knows what this prompt even does.

Anyhow, batch files are bad. So, let's replace them with something
better: a .exe shim, which doesn't have any of these issues!

This is what this PR does.